Initialize file name instead of path for FW API
When using standard firmware API to load firmware, it has to use the
name of the firmware file instead of the path.  The fw_path is set
to the path in dhd_update_fw_nv_path() wrongly and leads firmware
loading failure.

Initialize firmware_path, nvram_path and clm_path by NAME to fix it.

Bug: 188735328
Test: Enable WiFi and connect to AP successfully
Signed-off-by: Stephen Chu <stephen.chu@synaptics.corp-partner.google.com>
Change-Id: I248a1f6dfd6b62a1ff3e74b8d029b694edc87e9b