dhd: Fixed S2MPU fault due to wrong DMA direction
after S2MPU code update, the S2MPU fault detected as below :

08-06 21:23:09.222  root     0     0 E         : [    C2] s2mpu 14480000.s2mpu: IRQ (252): FAULT_STATUS=0x2
08-06 21:23:09.222  root     0     0 E         : [    C2] PA=0x0x00000000a983f0d8
08-06 21:23:09.222  root     0     0 E         : [    C2] DIRECTION=read
08-06 21:23:09.222  root     0     0 E         : [    C2] FAULT_TYPE=access perm fault
08-06 21:23:09.222  root     0     0 E         : [    C2] FAULT_INFO=0x1480000
08-06 21:23:19.938  root   678   678 I sscoredump sscd_wlan: crash: Dongle_Trap_pciedev_dpc+0x237__pciedev_dpc+0xb

By checking the dump, we could find the erroneous host address in Rx DMA.
To fix the problem, we change Rxbufferpost DMA direction to BIDIRECTIONAL.

The reason of the bidirectional operation is :
there are certain Rx data frames which are processed in the dongle (not to be consumed by the host. For example TDLS, Neighbour Discovery(IPV6) offload, etc),
since these are data packets, they will be received in the host rx buffer by default. As part of processing these frames in firmware, dongle identifies them, schedules pktfetch & reads the entire host rx buffer to fetch the payload to dongle for processing them further.

Bug: 195783356
Test: Run and pass traffic, SSR issue is no loner reproducible
Signed-off-by: Victor Hsu <hsuvictor@google.com>
Change-Id: Id2ce6eba673291396eb1e3306a2bba57af002e1b