DHD: Check validity of dhd pointer in runtime bus wake
When dhd_attach fails to allocate memory, the dhd_pub
pointer is not assigned to bus struct yet.

The failure handling calls dhd_detach for cleanup, which
stops rpm timer but no check of dhd pointer in
dhd_runtime_bus_wake. Check the NULL pointer to handle
the case correctly.

Bug: 215006470
Test: Wi-Fi on/off
Signed-off-by: Stephen Chu <stephen.chu@synaptics.corp-partner.google.com>
Change-Id: I995286da23ce724a37e6c42053e30f6af4a10776