DHD: Stop tx if DHD is still is suspended state
Log flooding as below:

[238181.664205][    C0] dhd_start_xmit: bus is in suspend(3) or
suspending(0x1) state!!
...

The dhd_bus_stop_queue shall be called but it was not due to the state
is not "suspend in process". However, we shall also need to stop tx if
the bus is suspended and not "resume in process".

If we don't stop the netif tx, the netif tx will run into infinite loop
because we always return NETDEV_TX_BUSY. NETDEV_TX_BUSY will cause the
netif tx task retry to call dhd_start_xmit forever.

Add condition checking to stop the tx queue if bus is suspend and
resume is not in process.

Bug: 227233684
Test:
Local test with Google tablet C6T6 by doing run-time suspend/resume
Signed-off-by: Stephen Chu <stephen.chu@synaptics.corp-partner.google.com>
Change-Id: I1968d50e454281353dc7fa77f0166b60cae154be