DHD: Disabled TCP SYNC flood protection
A connection failure happened with TCP sync flood notification.
The feature was a kind of WAR because old kernel didn't have any protection for TCP SYNC flood attack at that moment.
Currently the kernel has the protection itself with iptables and we don't use this feature anymore (from Android Q),
so let's disable this feature.

Bug: 269515906
Test: Validated basic test.
Change-Id: I714968b50ae2c6208d16d7adc77533eb86ce5ffb