DHD: Defer wl_scan_timeout timer if D3 acked
We want to collect the memdump when wl_scan_timeout happens.
But if it happens after D3 acked, the memdump will be
aborted. So we defer the wl_scan_timeout timer if D3 acked
to make sure the wl_scan_timeout will not run in D3 state.

Bug: 265581457
Test:
Verified with T6P to force wl_scan_timeout after D3 acked
Change-Id: Iddc7eaeb3b76d27c846f7180a810dec49e0b3515
Signed-off-by: Stephen Chu <stephen.chu@synaptics.corp-partner.google.com>